% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loopr.R
\name{amend}
\alias{amend}
\title{Amend a dataframe with new information}
\usage{
amend(data, amendData, by = NULL, suffix = "toFix")
}
\arguments{
\item{data}{A data frame}

\item{amendData}{A data frame}

\item{by}{A quoted vector of column names to join by. If set to NULL or unspecified, will default to the grouping columns in data}

\item{suffix}{A suffix used internally. No existing column names should use this suffix.}
}
\value{
An amended \code{\link{tbl_df}}
}
\description{
\code{\link{full_join}} two dataframes. If there are matching columns,
amend each \code{data} column with the corresponding \code{amendData} column using \code{\link{amendColumns}}.
}

