% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loopr.R
\name{amendColumns}
\alias{amendColumns}
\title{Amend variables with new information}
\usage{
amendColumns(data, originalNames, amendNames)
}
\arguments{
\item{data}{A data frame}

\item{originalNames}{A vector of column names with out-of-date information}

\item{amendNames}{A vector of column names with amended information. They will be removed at the end of processing.}
}
\value{
An amended \code{\link{tbl_df}}
}
\description{
Replace all non-NA values in one set of columns with values from another matching set
}

