% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loopr.R
\name{fillColumns}
\alias{fillColumns}
\title{Fill variables with new information}
\usage{
fillColumns(data, originalNames, fillNames)
}
\arguments{
\item{data}{A data frame}

\item{originalNames}{A vector of column names with out-of-date information}

\item{fillNames}{A vector of column names with new information. They will be removed at the end of processing.}
}
\value{
A \code{\link{tbl_df}}
}
\description{
Replace all NA values in one set of columns with values from another matching set
}

