% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loopr.R
\name{insert}
\alias{insert}
\title{Insert new information into a dataframe.}
\usage{
insert(data, insertData, by)
}
\arguments{
\item{data}{A data frame}

\item{insertData}{A data frame}

\item{by}{A quoted vector of column names to join by.}
}
\value{
An inserted \code{\link{tbl_df}}
}
\description{
\code{\link{anti_join}} data with insertData, then \code{\link{bind_cols}} of insertData, then arrange by \code{by} variables.
}

