% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loopr.R
\docType{class}
\name{loopClass}
\alias{loopClass}
\title{An implementation of an loop}
\format{An \code{\link{R6Class}} generator object}
\usage{
loopClass
}
\description{
An implementation of an loop
}
\section{Inherits}{

\describe{\item{\code{\link{stackClass}}}{}}
}

\section{Methods}{

\describe{\item{\code{
  begin(item, name = "")}}{
    Alias for \code{\link{stackClass}$push}}}

\describe{\item{\code{
  end(endData, FUN, ...)}}{
    Will return \code{FUN(\link{stackClass}$pop, endData, ...)}}}

\describe{\item{\code{
  cross(crossData, FUN, ...)}}{
    Will return \code{FUN(crossData, \link{stackClass}$pop, ...)}}}
}
\keyword{data}

