\name{loopr-package}
\alias{loopr-package}
\alias{loopr}
\docType{package}
\title{
Uses an archive to amend previous stages of a pipe using current output.
}
\description{
This methodology allows programmers to focus on only one aspect of data, and reincorporate changes to the whole.
}
\details{
\tabular{ll}{
Package: \tab loopr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-04\cr
License: \tab CC0\cr
}
}
\author{
Brandon Taylor

Maintainer: Brandon Taylor <Brandon.Taylor221@gmail.com>
}