% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loopr.R
\docType{class}
\name{stackClass}
\alias{stackClass}
\title{An implementation of a stack}
\format{An \code{\link{R6Class}} generator object}
\usage{
stackClass
}
\description{
An implementation of a stack
}
\section{Fields}{

\describe{
  \item{\code{stack}}{A list which can directly accessed or accessed through the functions below. The bottom (original) item, set to NULL, should not be modified.}}
}

\section{Methods}{

\describe{
  \item{\code{push(item, name = "")}}{Append an \code{item} to the \code{stack} with an optional \code{name} and return the item}}
}

\section{Active Bindings}{

\describe{
  \item{\code{height}}{This will return the length of the \code{stack}}}

\describe{
  \item{\code{pop}}{This will remove the last (most recent) item from the \code{stack} and return it}}

\describe{
  \item{\code{pop}}{This will return, but not remove, the last (most recent) item from the \code{stack}}}
}
\keyword{datasets}

