% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loopr.R
\docType{class}
\name{loop}
\alias{loop}
\title{An implementation of an loop}
\format{An \code{\link{R6Class}} generator object}
\usage{
loop
}
\description{
An implementation of an loop
}
\section{Inherits}{

\describe{\item{\code{\link{stack}}}{}}
}

\section{Methods}{

\describe{\item{\code{
  begin(item, name = "")}}{
    Alias for \code{\link{stack}$push}}}

\describe{\item{\code{
  end(endData, FUN, ...)}}{
    Will return \code{FUN(\link{stack}$pop, endData, ...)}}}

\describe{\item{\code{
  cross(crossData, FUN, ...)}}{
    Will return \code{FUN(crossData, \link{stack}$pop, ...)}}}
}
\keyword{data}

